+++ lib/samples/onboarding/onboarding_model_sample.dart
@@ -1,0 +1,17 @@
+String onboardingModelSample() {
+  return '''
+import 'package:flutter/material.dart';
+
+class OnboardingModel {
+  final String title;
+  final String description;
+  final IconData icon;
+
+  OnboardingModel({
+    required this.title,
+    required this.description,
+    required this.icon,
+  });
+}
+''';
+}